refactor: extract traits, class features, accordion modules
