+++ js/script.js
@@ -8,9 +8,17 @@ import { handleVariantExtraSelections, handleVariantFeatureChoices } from './var
 import { handleExtraLanguages, handleExtraSkills, handleExtraTools, handleExtraAncestry, gatherRaceTraitSelections } from './extrasSelections.js';
 import { openExtrasModal, updateExtraSelectionsView, showExtraSelection, extraCategoryAliases, extraCategoryDescriptions } from './extrasModal.js';
 import { setExtraSelections } from './extrasState.js';
+import { displayRaceTraits } from './raceTraits.js';
+import { updateSubclasses, renderClassFeatures } from './classFeatures.js';
+import { convertDetailsToAccordion, initializeAccordion } from './ui/accordion.js';
 
 
 let selectedData = getSelectedData();
+
+export function refreshSelectedData() {
+  selectedData = getSelectedData();
+  return selectedData;
+}
 
 export function checkTraitCompletion(detailId) {
   const detail = document.getElementById(detailId);
@@ -238,7 +246,7 @@ function initFeatureSelectionHandlers(container, saveCallback) {
   });
 }
 
-function saveFeatureSelection(select) {
+export function saveFeatureSelection(select) {
   const feature = select.dataset.feature;
   const index = select.dataset.index || 0;
   if (!feature) return;
@@ -247,667 +255,12 @@ function saveFeatureSelection(select) {
   saveSelectedData();
 }
 
-function convertDetailsToAccordion(container) {
-  if (!container) return;
-  container.querySelectorAll('details').forEach(det => {
-    const summary = det.querySelector('summary');
-    const contentNodes = Array.from(det.childNodes).filter(n => n !== summary);
-    const item = document.createElement('div');
-    const classes = ['accordion-item', 'feature-block', ...det.classList];
-    item.className = classes.join(' ');
-    if (det.id) item.id = det.id;
-
-    const header = document.createElement('button');
-    header.type = 'button';
-    header.className = 'accordion-header';
-    header.innerHTML = summary ? summary.innerHTML : '';
-    item.appendChild(header);
-
-    const content = document.createElement('div');
-    content.className = 'accordion-content';
-    contentNodes.forEach(n => content.appendChild(n));
-    item.appendChild(content);
-
-    det.replaceWith(item);
-  });
-}
-
-function initializeAccordion(root) {
-  if (!root) return;
-
-  root.querySelectorAll('.accordion-header').forEach((header, index) => {
-    const content = header.nextElementSibling;
-
-    // Apply ARIA roles and relationships
-    header.setAttribute('role', 'button');
-    if (!header.hasAttribute('tabindex')) header.setAttribute('tabindex', '0');
-    header.setAttribute('aria-expanded', 'false');
-
-    if (content) {
-      const headerId = header.id || `accordion-header-${index}`;
-      header.id = headerId;
-      content.setAttribute('role', 'region');
-      content.setAttribute('aria-labelledby', headerId);
-    }
-
-    const toggleAccordion = () => {
-      const expanded = header.getAttribute('aria-expanded') === 'true';
-      header.setAttribute('aria-expanded', String(!expanded));
-      header.classList.toggle('active', !expanded);
-      if (content) {
-        content.classList.toggle('show', !expanded);
-      }
-      header.focus();
-    };
-
-    header.addEventListener('click', toggleAccordion);
-    header.addEventListener('keydown', e => {
-      if (e.key === 'Enter' || e.key === ' ') {
-        e.preventDefault();
-        toggleAccordion();
-      }
-    });
-  });
-}
-
 /**
  * Render extra selections inside an accordion rather than a modal.
  */
 
 updateExtraSelectionsView();
 
-document.getElementById("raceSelect").addEventListener("change", () => {
-  console.log("🔄 Razza cambiata, reset delle selezioni extra...");
-  setSelectedData({}); // Reset delle selezioni extra
-  selectedData = getSelectedData();
-  document.getElementById("languageSelection").innerHTML = "";
-  document.getElementById("skillSelectionContainer").innerHTML = "";
-  document.getElementById("toolSelectionContainer").innerHTML = "";
-
-  displayRaceTraits(); // Ricarica i tratti della nuova razza
-  document.getElementById("confirmRaceSelection").style.display = "inline-block";
-});
-
-
-// ==================== DISPLAY DEI TRATTI DELLA RAZZA ====================
-async function displayRaceTraits() {
-  console.log("🛠 Esecuzione displayRaceTraits()...");
-  const racePath = document.getElementById("raceSelect").value;
-  const raceTraitsDiv = document.getElementById("raceTraits");
-  const racialBonusDiv = document.getElementById("racialBonusSelection");
-
-  if (!racePath) {
-    console.warn("⚠️ displayRaceTraits(): Nessuna razza selezionata.");
-    raceTraitsDiv.textContent = '';
-    raceTraitsDiv.appendChild(createParagraph('Seleziona una razza per vedere i tratti.'));
-    racialBonusDiv.style.display = "none";
-    resetRacialBonuses();
-    return;
-  }
-
-  console.log(`📜 Caricamento tratti per ${racePath}...`);
-  try {
-    const response = await fetch(racePath);
-    if (!response.ok) throw new Error('Network response was not ok');
-    const data = await response.json();
-    console.log("📜 Dati razza caricati:", data);
-    const raceData = convertRaceData(data);
-    raceTraitsDiv.textContent = '';
-    raceTraitsDiv.appendChild(createHeader(`Tratti di ${raceData.name}`, 3));
-
-      // Speed
-      let speedText = 'Velocità: Non disponibile';
-      if (raceData.speed) {
-        if (typeof raceData.speed === 'object') {
-          const parts = Object.entries(raceData.speed).map(([t, v]) => `${t}: ${v} ft`);
-          speedText = `Velocità: ${parts.join(', ')}`;
-        } else {
-          speedText = `Velocità: ${raceData.speed} ft`;
-        }
-      }
-      raceTraitsDiv.appendChild(createParagraph(speedText));
-
-      // Darkvision
-      if (raceData.senses && raceData.senses.darkvision) {
-        raceTraitsDiv.appendChild(createParagraph(`Visione: ${raceData.senses.darkvision} ft`));
-      }
-
-      // Trait details
-      let cantripDetail = null;
-      if (raceData.traits && raceData.traits.length > 0) {
-        raceTraitsDiv.appendChild(createHeader('Tratti:', 4));
-        raceData.traits.forEach(trait => {
-          const traitId = `trait-${trait.name.toLowerCase().replace(/[^a-z0-9]+/g, '-')}`;
-          const featureKey = trait.name;
-          const details = document.createElement('details');
-          details.className = 'race-trait feature-block';
-          details.id = traitId;
-          const summary = document.createElement('summary');
-          summary.textContent = trait.name;
-          details.appendChild(summary);
-          const desc = document.createElement('div');
-          desc.className = 'feature-desc';
-          desc.textContent = trait.description || '';
-          details.appendChild(desc);
-          const choices = trait.choices || trait.variant_feature_choices;
-          if (choices) {
-            choices.forEach((choice, cIdx) => {
-              const options = choice.options || choice.selection || [];
-              const saved = selectedData[featureKey]?.[cIdx] || '';
-              const label = document.createElement('label');
-              label.textContent = (choice.name || 'Scegli') + ': ';
-              const select = document.createElement('select');
-              select.dataset.feature = featureKey;
-              select.dataset.index = cIdx;
-              const defaultOpt = document.createElement('option');
-              defaultOpt.value = '';
-              defaultOpt.textContent = 'Seleziona...';
-              select.appendChild(defaultOpt);
-              options.forEach(opt => {
-                const option = document.createElement('option');
-                option.value = opt;
-                option.textContent = opt;
-                if (saved === opt) option.selected = true;
-                select.appendChild(option);
-              });
-              label.appendChild(select);
-              details.appendChild(label);
-            });
-          }
-          if (trait.name.toLowerCase().includes('cantrip')) {
-            cantripDetail = details;
-          }
-          raceTraitsDiv.appendChild(details);
-        });
-      }
-
-      // Extra selections (languages, skills, tools) merged into existing traits
-      const extraSelections = gatherExtraSelections(raceData, 'race');
-      setExtraSelections(extraSelections);
-      const detailMatchers = {
-        'Languages': /language/i,
-        'Skill Proficiency': /skill/i,
-        'Tool Proficiency': /tool/i
-      };
-      extraSelections.forEach((choice, idx) => {
-        const featureKey = extraCategoryAliases[choice.name] || choice.name;
-        let targetDetail = null;
-        const matcher = detailMatchers[choice.name];
-        if (matcher) {
-          targetDetail = Array.from(raceTraitsDiv.querySelectorAll('details'))
-            .find(det => matcher.test(det.querySelector('summary')?.textContent || ''));
-        }
-        if (!targetDetail) {
-          targetDetail = document.createElement('details');
-          targetDetail.className = 'race-trait feature-block';
-          targetDetail.id = `race-extra-${idx}`;
-          const summary = document.createElement('summary');
-          summary.textContent = choice.name;
-          targetDetail.appendChild(summary);
-          raceTraitsDiv.appendChild(targetDetail);
-        }
-        if (choice.description && !targetDetail.querySelector('.feature-desc')) {
-          const desc = document.createElement('div');
-          desc.className = 'feature-desc';
-          desc.textContent = choice.description;
-          targetDetail.appendChild(desc);
-        }
-        const options = choice.selection || [];
-        const count = choice.count || 1;
-        for (let i = 0; i < count; i++) {
-          const saved = selectedData[featureKey]?.[i] || '';
-          const label = document.createElement('label');
-          label.textContent = `${choice.name}${count > 1 ? ' ' + (i + 1) : ''}: `;
-          const select = document.createElement('select');
-          select.dataset.feature = featureKey;
-          select.dataset.index = i;
-          const def = document.createElement('option');
-          def.value = '';
-          def.textContent = 'Seleziona...';
-          select.appendChild(def);
-          options.forEach(opt => {
-            const option = document.createElement('option');
-            option.value = opt;
-            option.textContent = opt;
-            if (saved === opt) option.selected = true;
-            select.appendChild(option);
-          });
-          label.appendChild(select);
-          targetDetail.appendChild(label);
-        }
-      });
-
-      // Variant feature choices
-      if (raceData.variant_feature_choices) {
-        const details = document.createElement('details');
-        details.className = 'race-trait feature-block needs-selection incomplete';
-        details.id = 'variantFeatureTrait';
-        const summary = document.createElement('summary');
-        summary.textContent = 'Variant Feature';
-        details.appendChild(summary);
-        const selContainer = document.createElement('div');
-        selContainer.id = 'variantFeatureSelectionContainer';
-        details.appendChild(selContainer);
-        const extraContainer = document.createElement('div');
-        extraContainer.id = 'variantExtraContainer';
-        details.appendChild(extraContainer);
-        raceTraitsDiv.appendChild(details);
-      }
-
-      // Ancestry placeholder
-      const ancestryDetails = document.createElement('details');
-      ancestryDetails.className = 'race-trait feature-block needs-selection incomplete hidden';
-      ancestryDetails.id = 'ancestryTrait';
-      const ancestrySummary = document.createElement('summary');
-      ancestrySummary.textContent = 'Ancestry';
-      ancestryDetails.appendChild(ancestrySummary);
-      const ancestryContainer = document.createElement('div');
-      ancestryContainer.id = 'ancestrySelectionContainer';
-      ancestryDetails.appendChild(ancestryContainer);
-      raceTraitsDiv.appendChild(ancestryDetails);
-
-      // Tables (rawEntries)
-      const tablesHtml = renderTables(raceData.rawEntries);
-      if (tablesHtml) {
-        const wrapper = document.createElement('div');
-        wrapper.appendChild(document.createRange().createContextualFragment(tablesHtml));
-        raceTraitsDiv.appendChild(wrapper);
-      }
-
-      await handleSpellcasting(raceData, cantripDetail);
-      handleVariantFeatureChoices(raceData);
-      handleExtraAncestry(raceData, 'ancestrySelectionContainer');
-
-      convertDetailsToAccordion(raceTraitsDiv);
-      raceTraitsDiv.classList.add('accordion');
-      raceTraitsDiv.style.display = 'block';
-      initFeatureSelectionHandlers(raceTraitsDiv, saveFeatureSelection);
-      initializeAccordion(raceTraitsDiv);
-
-      const ancestryDetail = document.getElementById('ancestryTrait');
-      if (ancestryDetail) {
-        const ancContainer = document.getElementById('ancestrySelectionContainer');
-        if (ancContainer && ancContainer.innerHTML.trim() !== '') {
-          ancestryDetail.classList.remove('hidden');
-        }
-      }
-
-      const variantSelect = document.getElementById('variantFeatureChoice');
-      if (variantSelect) {
-        variantSelect.addEventListener('change', () => {
-          selectedData['Variant Feature'] = [variantSelect.value];
-          handleVariantExtraSelections();
-          checkTraitCompletion('variantFeatureTrait');
-        });
-        checkTraitCompletion('variantFeatureTrait');
-      }
-
-      const ancestrySelect = document.getElementById('ancestrySelection');
-      if (ancestrySelect) {
-        ancestrySelect.addEventListener('change', () => {
-          selectedData['Ancestry'] = [ancestrySelect.value];
-          checkTraitCompletion('ancestryTrait');
-        });
-        checkTraitCompletion('ancestryTrait');
-      }
-
-      resetRacialBonuses();
-      window.currentRaceData = raceData;
-  } catch (error) {
-    handleError(`Errore caricando i tratti della razza: ${error}`);
-  }
-}
-
-// ==================== UPDATE SUBCLASSES (STEP 5) ====================
-async function updateSubclasses() {
-  const classPath = document.getElementById("classSelect").value;
-  const featuresDiv = document.getElementById("classFeatures");
-  if (!classPath) {
-    window.currentClassData = null;
-    if (featuresDiv) {
-      featuresDiv.textContent = '';
-      featuresDiv.appendChild(createParagraph('Seleziona una classe per vedere i tratti.'));
-    }
-    return;
-  }
-  try {
-    const response = await fetch(classPath);
-    if (!response.ok) throw new Error('Network response was not ok');
-    const data = await response.json();
-    window.currentClassData = data;
-    renderClassFeatures();
-  } catch (error) {
-    handleError(`Errore caricando le sottoclasse: ${error}`);
-  }
-}
-
-function getSubclassFilename(name) {
-  let lower = name.toLowerCase();
-  if (lower.startsWith("college of ")) {
-    lower = "college_of_" + lower.slice("college of ".length);
-  } else if (lower.startsWith("circle of the ")) {
-    lower = "circle_of_the_" + lower.slice("circle of the ".length);
-  } else if (lower.startsWith("circle of ")) {
-    lower = "circle_of_" + lower.slice("circle of ".length);
-  } else {
-    lower = lower
-      .replace(/^(path|oath|way|school) of the /, "")
-      .replace(/^(path|oath|way|school) of /, "")
-      .replace(/^the /, "");
-    if (lower.endsWith(" domain")) {
-      lower = lower.slice(0, -" domain".length);
-    }
-  }
-  let file = lower.replace(/ /g, "_") + ".json";
-  if (file === "wild_magic.json" && window.currentClassData?.name === "Barbarian") {
-    file = "barbarian_wild_magic.json";
-  }
-  return file;
-}
-
-async function renderClassFeatures() {
-  const featuresDiv = document.getElementById('classFeatures');
-  const data = window.currentClassData;
-  if (!featuresDiv) return;
-  if (!data) {
-    featuresDiv.textContent = '';
-    featuresDiv.appendChild(createParagraph('Seleziona una classe per vedere i tratti.'));
-    return;
-  }
-  const charLevel = parseInt(document.getElementById('levelSelect')?.value) || 1;
-  const subclassName = document.getElementById('subclassSelect')?.value || '';
-  featuresDiv.textContent = '';
-  featuresDiv.appendChild(createHeader(data.name, 3));
-  if (data.description) featuresDiv.appendChild(createParagraph(data.description));
-
-  // Hit points section
-  if (data.hit_die || data.hp_at_1st_level || data.hp_at_higher_levels) {
-    const details = document.createElement('details');
-    details.className = 'feature-block';
-    const summary = document.createElement('summary');
-    summary.textContent = 'Hit Points';
-    details.appendChild(summary);
-    if (data.hit_die) details.appendChild(createParagraph(`Hit Die: ${data.hit_die}`));
-    if (data.hp_at_1st_level) details.appendChild(createParagraph(`Hit Points at 1st Level: ${data.hp_at_1st_level}`));
-    if (data.hp_at_higher_levels) details.appendChild(createParagraph(`Hit Points at Higher Levels: ${data.hp_at_higher_levels}`));
-    featuresDiv.appendChild(details);
-  }
-
-  // Proficiencies section
-  const profTexts = [];
-  let toolChoice = null;
-  if (data.saving_throws) profTexts.push(`Saving Throw Proficiencies: ${data.saving_throws.join(', ')}`);
-  if (data.weapon_proficiencies) profTexts.push(`Weapon Proficiencies: ${data.weapon_proficiencies.join(', ')}`);
-  if (data.tool_proficiencies) {
-    if (Array.isArray(data.tool_proficiencies)) {
-      const fixed = data.tool_proficiencies.filter(tp => {
-        const lower = tp.toLowerCase();
-        return !lower.includes('of your choice') && !lower.includes(' or ');
-      });
-      if (fixed.length) profTexts.push(`Tool Proficiencies: ${fixed.join(', ')}`);
-      data.tool_proficiencies.forEach(tp => {
-        const lower = tp.toLowerCase();
-        const wordMap = { one: 1, two: 2, three: 3 };
-        let count = 1;
-        for (const word in wordMap) {
-          if (lower.includes(word)) { count = wordMap[word]; break; }
-        }
-        if (lower.includes("artisan's tools") && lower.includes('musical instrument')) {
-          toolChoice = { choose: count, options: [...ARTISAN_TOOLS, ...MUSICAL_INSTRUMENTS] };
-        } else if (lower.includes('musical instrument')) {
-          toolChoice = { choose: count, options: MUSICAL_INSTRUMENTS };
-        } else if (lower.includes("artisan's tools")) {
-          toolChoice = { choose: count, options: ARTISAN_TOOLS };
-        }
-      });
-      if (toolChoice) {
-        profTexts.push(`Tool Proficiencies: Choose ${toolChoice.choose} from ${toolChoice.options.join(', ')}`);
-      }
-    } else if (data.tool_proficiencies.options) {
-      const fixed = data.tool_proficiencies.fixed || [];
-      if (fixed.length) profTexts.push(`Tool Proficiencies: ${fixed.join(', ')}`);
-      if (data.tool_proficiencies.choose && data.tool_proficiencies.options.length) {
-        toolChoice = {
-          choose: data.tool_proficiencies.choose,
-          options: data.tool_proficiencies.options
-        };
-        profTexts.push(`Tool Proficiencies: Choose ${toolChoice.choose} from ${toolChoice.options.join(', ')}`);
-      }
-    }
-  }
-  if (data.armor_proficiencies) {
-    profTexts.push(`Armor Training: ${data.armor_proficiencies.join(', ')}`);
-  }
-  if (data.skill_proficiencies && data.skill_proficiencies.choose) {
-    profTexts.push(`Skill Proficiencies: Choose ${data.skill_proficiencies.choose} from ${data.skill_proficiencies.options.join(', ')}`);
-  }
-  if (data.multiclassing && data.multiclassing.prerequisites) {
-    const prereqs = Object.entries(data.multiclassing.prerequisites)
-      .map(([ability, score]) => `${ability} ${score}`)
-      .join(', ');
-    profTexts.push(`Multiclassing Prerequisite: ${prereqs}`);
-  }
-  if (profTexts.length) {
-    const details = document.createElement('details');
-    details.className = 'feature-block';
-    const summary = document.createElement('summary');
-    summary.textContent = 'Proficiencies';
-    details.appendChild(summary);
-    profTexts.forEach(t => details.appendChild(createParagraph(t)));
-    featuresDiv.appendChild(details);
-  }
-
-  // Subclasses select
-  if (data.subclasses && data.subclasses.length > 0) {
-    const details = document.createElement('details');
-    details.className = 'feature-block';
-    details.id = 'subclassTrait';
-    const summary = document.createElement('summary');
-    summary.textContent = 'Sottoclasse';
-    details.appendChild(summary);
-    const select = document.createElement('select');
-    select.id = 'subclassSelect';
-    select.className = 'form-control';
-    const def = document.createElement('option');
-    def.value = '';
-    def.textContent = 'Seleziona una sottoclasse';
-    select.appendChild(def);
-    data.subclasses.forEach(sc => {
-      const option = document.createElement('option');
-      option.value = sc.name;
-      option.textContent = sc.name;
-      if (subclassName === sc.name) option.selected = true;
-      select.appendChild(option);
-    });
-    details.appendChild(select);
-    featuresDiv.appendChild(details);
-  }
-
-  // Fetch subclass data
-  let subData = null;
-  if (subclassName) {
-    try {
-      const file = getSubclassFilename(subclassName);
-      const resp = await fetch(`data/subclasses/${file}`);
-      if (!resp.ok) throw new Error('Network response was not ok');
-      subData = await resp.json();
-      if (subData.description) {
-        featuresDiv.appendChild(createParagraph(subData.description));
-      }
-    } catch (err) {
-      featuresDiv.appendChild(createParagraph('Dettagli della sottoclasse non disponibili.'));
-    }
-  } else {
-    featuresDiv.appendChild(createParagraph('Seleziona una sottoclasse per vedere i tratti.'));
-  }
-
-  // Merge class and subclass features by level
-  const mergedFeatures = {};
-  if (data.features_by_level) {
-    Object.entries(data.features_by_level).forEach(([lvl, feats]) => {
-      mergedFeatures[lvl] = [...feats];
-    });
-  }
-  if (subData?.features_by_level) {
-    Object.entries(subData.features_by_level).forEach(([lvl, feats]) => {
-      if (!mergedFeatures[lvl]) mergedFeatures[lvl] = [];
-      mergedFeatures[lvl].push(...feats);
-    });
-  }
-  const levels = Object.keys(mergedFeatures).sort((a, b) => a - b);
-  levels.forEach(lvl => {
-    if (parseInt(lvl) <= charLevel) {
-      featuresDiv.appendChild(createHeader(`Livello ${lvl}`, 4));
-      mergedFeatures[lvl].forEach((f, idx) => {
-        if (typeof f === 'string') {
-          const details = document.createElement('details');
-          details.className = 'feature-block';
-          const summary = document.createElement('summary');
-          summary.textContent = f;
-          details.appendChild(summary);
-          featuresDiv.appendChild(details);
-        } else {
-          const details = document.createElement('details');
-          details.className = 'feature-block';
-          details.id = `class-feature-${lvl}-${idx}`;
-          const summary = document.createElement('summary');
-          summary.textContent = f.name;
-          details.appendChild(summary);
-          if (f.description) {
-            const desc = document.createElement('div');
-            desc.className = 'feature-desc';
-            desc.textContent = f.description;
-            details.appendChild(desc);
-          }
-          const choices = f.choices || f.variant_feature_choices;
-          if (choices) {
-            choices.forEach((choice, cIdx) => {
-              const options = choice.options || choice.selection || [];
-              const saved = selectedData[f.name]?.[cIdx] || '';
-              const label = document.createElement('label');
-              label.textContent = (choice.name || 'Scegli') + ': ';
-              const select = document.createElement('select');
-              select.dataset.feature = f.name;
-              select.dataset.index = cIdx;
-              const optDef = document.createElement('option');
-              optDef.value = '';
-              optDef.textContent = 'Seleziona...';
-              select.appendChild(optDef);
-              options.forEach(opt => {
-                const option = document.createElement('option');
-                option.value = opt;
-                option.textContent = opt;
-                if (saved === opt) option.selected = true;
-                select.appendChild(option);
-              });
-              label.appendChild(select);
-              details.appendChild(label);
-            });
-          }
-          featuresDiv.appendChild(details);
-        }
-      });
-    }
-  });
-
-  const allChoices = [...(data.choices || []), ...(subData?.choices || [])];
-  if (toolChoice && !allChoices.some(c => c.name === 'Tool Proficiency')) {
-    allChoices.push({
-      level: 1,
-      name: 'Tool Proficiency',
-      description: `Choose ${toolChoice.choose} from the class options`,
-      count: toolChoice.choose,
-      selection: toolChoice.options
-    });
-  }
-  if (data.skill_proficiencies && !allChoices.some(c => c.name === 'Skill Proficiency')) {
-    allChoices.push({
-      level: 1,
-      name: 'Skill Proficiency',
-      description: `Choose ${data.skill_proficiencies.choose} from the class skill list`,
-      count: data.skill_proficiencies.choose,
-      selection: data.skill_proficiencies.options
-    });
-  }
-  const extraSelections = gatherExtraSelections({ choices: allChoices }, 'class', charLevel);
-  setExtraSelections(extraSelections);
-  const detailMatchers = {
-    'Languages': /Proficiencies/i,
-    'Skill Proficiency': /Proficiencies/i,
-    'Tool Proficiency': /Proficiencies/i
-  };
-  extraSelections.forEach(choice => {
-    const featureKey = extraCategoryAliases[choice.name] || choice.name;
-    const allDetails = Array.from(featuresDiv.querySelectorAll('details'));
-    let targetDetail = allDetails.find(det => {
-      const txt = det.querySelector('summary')?.textContent || '';
-      return txt === choice.name || txt === featureKey;
-    });
-    if (!targetDetail) {
-      const matcher = detailMatchers[choice.name] || detailMatchers[featureKey];
-      if (matcher) {
-        targetDetail = allDetails.find(det => matcher.test(det.querySelector('summary')?.textContent || ''));
-      }
-    }
-    if (!targetDetail) {
-      targetDetail = document.createElement('details');
-      targetDetail.className = 'feature-block';
-      const summary = document.createElement('summary');
-      summary.textContent = choice.name;
-      targetDetail.appendChild(summary);
-      featuresDiv.appendChild(targetDetail);
-    }
-    if (choice.description && !targetDetail.querySelector('.feature-desc')) {
-      const desc = document.createElement('div');
-      desc.className = 'feature-desc';
-      desc.textContent = choice.description;
-      targetDetail.appendChild(desc);
-    }
-    const options = choice.options || choice.selection || [];
-    const selects = [];
-    for (let i = 0; i < (choice.count || 1); i++) {
-      const saved = choice.selected?.[i] || selectedData[featureKey]?.[i] || '';
-      const label = document.createElement('label');
-      label.textContent = `${choice.name}${choice.count > 1 ? ' ' + (i + 1) : ''}: `;
-      const select = document.createElement('select');
-      select.dataset.feature = featureKey;
-      select.dataset.index = i;
-      select.dataset.options = JSON.stringify(options);
-      const optDef = document.createElement('option');
-      optDef.value = '';
-      optDef.textContent = 'Seleziona...';
-      select.appendChild(optDef);
-      options.forEach(opt => {
-        const option = document.createElement('option');
-        option.value = opt;
-        option.textContent = opt;
-        select.appendChild(option);
-      });
-      if (saved) select.value = saved;
-      label.appendChild(select);
-      targetDetail.appendChild(label);
-      selects.push(select);
-    }
-
-    const update = () => {
-      const chosen = new Set(selects.map(s => s.value).filter(Boolean));
-      selects.forEach(sel => {
-        const opts = JSON.parse(sel.dataset.options);
-        const current = sel.value;
-        sel.innerHTML = `<option value="">Seleziona...</option>` +
-          opts.map(o => `<option value="${o}" ${chosen.has(o) && o !== current ? 'disabled' : ''}>${o}</option>`).join('');
-        sel.value = current;
-      });
-    };
-    selects.forEach(sel => sel.addEventListener('change', update));
-    update();
-  });
-
-  convertDetailsToAccordion(featuresDiv);
-  featuresDiv.classList.add('accordion');
-  initFeatureSelectionHandlers(featuresDiv, saveFeatureSelection);
-  initializeAccordion(featuresDiv);
-  const subSel = document.getElementById('subclassSelect');
   if (subSel) {
     subSel.addEventListener('change', renderClassFeatures);
   }
@@ -1124,7 +477,7 @@ function applyRacialBonuses() {
   updateFinalScores();
 }
 
-function resetRacialBonuses() {
+export function resetRacialBonuses() {
   document.getElementById("racialBonus1").value = "";
   document.getElementById("racialBonus2").value = "";
   document.getElementById("racialBonus3").value = "";
@@ -1156,6 +509,9 @@ export {
   initializeAccordion,
   updateSubclasses,
   renderClassFeatures,
+  saveFeatureSelection,
+  refreshSelectedData,
+  resetRacialBonuses,
   openExtrasModal,
   updateExtraSelectionsView,
   showExtraSelection,
